print command now can have spaces and tabs. Added Clean Code button. Added custom Data Cell. Added section tags. Added link to gitlab

+++ frontend/src/components/Context.js
@@ -9,6 +9,7 @@ const ContextNewCell = React.createContext()
 const ContextNewCellDown = React.createContext()
 const ContextDeleteCell = React.createContext()
 const ContextCopyToClipBoard = React.createContext()
+const ContextCleanCode = React.createContext()
 // const local_host = "http://localhost:8080"
 
 
@@ -43,6 +44,10 @@ export function useContextDeleteCell(){
 
 export function useContextCopyToClipBoard(){
     return useContext(ContextCopyToClipBoard)
+}
+
+export function useContextCleanCode(){
+    return useContext(ContextCleanCode)
 }
 
 
@@ -81,36 +86,52 @@ function Provider({children}){
         for(let i = 0; i < 10; i++){
             cellsData.push(initCell(i))
         }
-    
-        cellsData[0].code = ";data\n"
+
+        cellsData[0].code = ";Este programa se encarga de pasar los 2 primeros pixeles en memoria a escala de grises.\n"
+        cellsData[0].code += ";Esto se hace mediante la formula: gris = (rojo + 2 * verde + azul)/4.\n"
+        cellsData[0].code += ";Una vez se obtienen los 2 valores se les hace un broadcast de tal modo que "
+        cellsData[0].code += "la parte baja del registro este formada por el primer pixel gris repetido 2 veces y "
+        cellsData[0].code += "la parte alta contenga el segundo pixel gris.\n\n"
         cellsData[0].code += "pixeles: db 173, 68, 144, 255, 16, 54, 231, 255, 29, 178, 50, 255, 79, 211, 203, 255\n"
         cellsData[0].code += "mascara: db 0, 0, 0, 2, 0, 0, 0, 2, 1, 1, 1, 2, 1, 1, 1, 2"
     
-        cellsData[1].code = "movdqu xmm0, [mascara]\n"
+        cellsData[1].code = ";Cargo los pixeles en los registros xmm y los imprimo como enteros de 8 bits sin signo.\n"
+        cellsData[1].code += "movdqu xmm0, [mascara]\n"
         cellsData[1].code += "movdqu xmm1, [pixeles]\n"
-        cellsData[1].code += ";p xmm0.v16_int8\n"
+        cellsData[1].code += ";p/u xmm0.v16_int8\n"
         cellsData[1].code += ";p/u xmm1.v16_int8"
     
-        cellsData[2].code = "pmovzxbw xmm1, xmm1\n"
+        cellsData[2].code = ";Me quedo solo con la parte baja de xmm1 en forma de enteros de 16 bits e imprimo el registro de esa forma.\n"
+        cellsData[2].code += "pmovzxbw xmm1, xmm1\n"
         cellsData[2].code += ";p xmm1.v8_int16"
 
-        cellsData[3].code = "pshufhw xmm1, xmm1, 0b01100100\n"
+        cellsData[3].code = ";Reemplazo el valor alfa de cada pixel con el valor verde del mismo. Hacemos esto para ambos pixeles.\n"
+        cellsData[3].code += ";Esto nos permite conseguir el valor de gris haciendo sumas horizontales.\n"
+        cellsData[3].code += "pshufhw xmm1, xmm1, 0b01100100\n"
         cellsData[3].code += "pshuflw xmm1, xmm1, 0b01100100"
     
-        cellsData[4].code = "phaddw xmm1, xmm1"
+        cellsData[4].code = ";Hacemos la primer suma horizontal.\n"
+        cellsData[4].code += "phaddw xmm1, xmm1"
 
-        cellsData[5].code = "phaddw xmm1, xmm1"
+        cellsData[5].code = ";Hacemos la segunda suma horizontal.\n"
+        cellsData[5].code += ";Ahora tenemos en los 2 registros de la derecha el los valores de R + 2G + B de cada pixel.\n"
+        cellsData[5].code += "phaddw xmm1, xmm1"
 
-        cellsData[6].code = "psrlw xmm1, 3"
+        cellsData[6].code = ";Shifteamos a derecha para dividir los resultados por 4 y asi obtener el valor gris del pixel.\n"
+        cellsData[6].code += "psrlw xmm1, 2"
 
-        cellsData[7].code = "packuswb xmm1, xmm1\n"
+        cellsData[7].code = ";Volvemos a almacenar los pixeles conseguidos como enteros de 8 bits. En este puento vuelvo a imprimir en 8 bits.\n"
+        cellsData[7].code += "packuswb xmm1, xmm1\n"
         cellsData[7].code += ";p xmm1.v16_int8"
 
-        cellsData[8].code = "xor rax, rax\n"
+        cellsData[8].code = ";Inserto un 255 en el valor de alfa para restaurar el valor original.\n"
+        cellsData[8].code += "xor rax, rax\n"
         cellsData[8].code += "dec rax\n"
         cellsData[8].code += "pinsrb xmm1, al, 2"
 
-        cellsData[9].code = "pshufb xmm1, xmm0"
+        cellsData[9].code = ";Uso la mascara en xmm0 para hacer un broadcast de los valores conseguidos y del alfa "
+        cellsData[9].code += "como decia el enunciado.\n"
+        cellsData[9].code += "pshufb xmm1, xmm0"
 
         
 
@@ -162,9 +183,8 @@ function Provider({children}){
 
 
     function deleteCell(e, buttonNumber){
-        console.log(buttonNumber)
         e.preventDefault()
-        if(TotalCells > 1){
+        if(TotalCells > 2){
             let copy = JSON.parse(JSON.stringify(CellsData))//This is the only way of doing a depth copy
             copy.splice(buttonNumber, 1)
             copy = fixIndexing(copy)
@@ -208,6 +228,15 @@ function Provider({children}){
         }
 
         copyStringToClipboard(resText)
+        
+    }
+
+    function cleanCode(e){
+        if(window.confirm("Are you sure you want to delete ALL code?\nThere is no turning back.")){
+            e.preventDefault()
+            let copy = [initCell(0), initCell(1)]
+            setCellsData(copy)
+        }
         
     }
     function updateXMMData(data){
@@ -255,7 +284,9 @@ function Provider({children}){
                 newCell(event, parseInt(event.target.id))
             }
             if(event.key.toLowerCase() === "d" && event.ctrlKey && event.altKey){
-                deleteCell(event, parseInt(event.target.id))
+                if(parseInt(event.target.id) !== 0){//Won't delete data cell
+                    deleteCell(event, parseInt(event.target.id))
+                }
             }
         }
         else{
@@ -263,28 +294,29 @@ function Provider({children}){
                 newCell(event, VisualizerData.TotalCells)
             }
             if(event.key === "ArrowUp" && event.ctrlKey){
-                newCell(event, 0)
+                newCell(event, 1)
             }
         }
     }
-    // window.addEventListener("onkeypress", (event) => {
-        
-    // })
+
+   
 
     return(
         <ContextData.Provider value={VisualizerData}>
             <ContextSubmit.Provider value={submitCode}>
-                <ContextNewCell.Provider value={newCell}>
-                        <ContextDeleteCell.Provider value={deleteCell}>
-                            <ContextCopyToClipBoard.Provider value={copyToClipBoard}>
-                                <ContextUpdateData.Provider value={updateCodeData}>
-                                    <ContextXMMRegisters.Provider value={getXMMRegisters}>
-                                        {children}
-                                    </ContextXMMRegisters.Provider>
-                                </ContextUpdateData.Provider>
-                            </ContextCopyToClipBoard.Provider>
-                        </ContextDeleteCell.Provider>
-                </ContextNewCell.Provider>
+                <ContextCleanCode.Provider value={cleanCode}>
+                    <ContextNewCell.Provider value={newCell}>
+                            <ContextDeleteCell.Provider value={deleteCell}>
+                                <ContextCopyToClipBoard.Provider value={copyToClipBoard}>
+                                    <ContextUpdateData.Provider value={updateCodeData}>
+                                        <ContextXMMRegisters.Provider value={getXMMRegisters}>
+                                            {children}
+                                        </ContextXMMRegisters.Provider>
+                                    </ContextUpdateData.Provider>
+                                </ContextCopyToClipBoard.Provider>
+                            </ContextDeleteCell.Provider>
+                    </ContextNewCell.Provider>
+                </ContextCleanCode.Provider>
             </ContextSubmit.Provider>
         </ContextData.Provider>
     )
